Remove support for fully disabling tcache stats
The following commit added support for stripping the tcache stats field
from the cache struct when it's disabled:

commit a8b52518654725e64def5ebbd0ed472100e5a522
Author: Christopher Ferris <cferris@google.com>
Date:   Tue Apr 30 17:45:48 2019 -0700

    Remove unused stats from cache structure.

    Removing the stats make the whole cache structure fit in a single page.

    Bug: 131362671

    Test: Verified that all bionic malloc benchmarks are still the same.
    Test: It turns out that the malloc_sql benchmarks seem to get faster.
    Test: Verified that after this change, it saves about 2K PSS per thread.
    Change-Id: I4dcd633543f05f1a9d47db175f9977ddb42188a9

But jemalloc 5.3.0 adds a few more fields anyway, which invalidates the
original goal. Maintaining this across major jemalloc updates is too
much of a burden with little benefit at this point, so let's just ditch
it and rely on the compiler to optimize out stats accounting.

Change-Id: Ic4ef7f5c157b83439c555b98fca36757a6e83645